proj_iw/requirements.txt
asgiref==3.10.0
brotli==1.2.0
click==8.3.1
colorama==0.4.6
Django==5.2.7
gunicorn==23.0.0
h11==0.16.0
mysqlclient==2.2.7
packaging==25.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.38.0
whitenoise==6.11.0
